packages/sdk-py/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ainp-sdk"
version = "0.1.0a0"
description = "AINP Python SDK (alpha)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "AINP Contributors" }]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Development Status :: 3 - Alpha",
]

dependencies = [
  "pynacl>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/your-org/ainp"
